moved everything related with useReducer hook into useWhatIsThisGameReducer for a cleaner WhatIsThisGame component
